Update sponsor controller (#457)
* Update .gitignore to also exclude symlinks to build dir


* Add dataintegrityviolation handler for sponsor controller


* Add naming of foreing key constraint to get readable error


* Remove hashcode from sponsor


* Undo foreign key and work with custom exception


* Rename exception and replace find method with native implementation

+++ src/main/java/ch/wisv/areafiftylan/web/sponsor/service/SponsorServiceImpl.java
@@ -1,5 +1,6 @@
 package ch.wisv.areafiftylan.web.sponsor.service;
 
+import ch.wisv.areafiftylan.exception.SponsorConstraintViolationException;
 import ch.wisv.areafiftylan.web.sponsor.model.Sponsor;
 import ch.wisv.areafiftylan.web.sponsor.model.SponsorType;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -34,7 +35,12 @@ public class SponsorServiceImpl implements SponsorService {
 
     @Override
     public void deleteSponsor(Long sponsorId) {
-        sponsorRepository.delete(sponsorId);
+        Sponsor sponsor = sponsorRepository.findOne(sponsorId);
+        if (sponsor.getTournaments().size() == 0) {
+            sponsorRepository.delete(sponsorId);
+        } else {
+            throw new SponsorConstraintViolationException(sponsor);
+        }
     }
 
     @Override
